feat: ad findTop10

+++ src/main/kotlin/com/whoiswoony/springtutorial/dto/QuizSetResponse.kt
@@ -5,7 +5,7 @@ import com.querydsl.core.annotations.QueryProjection
 data class QuizSetResponse @QueryProjection constructor(
     val title:String,
     val description:String,
-    val shareNum: Long,
-    val viewNum: Long,
+    val shareNum: Int,
+    val viewNum: Int,
     val id:Long?=null,
 )
