+++ src/main/kotlin/com/whoiswoony/springtutorial/service/quizset/SharedQuizSetService.kt
@@ -40,8 +40,8 @@ class SharedQuizSetService (
                 val mySharedQuizSet = QuizSetResponse(
                     title = sharedQuizSet.quizSet.title,
                     description = sharedQuizSet.quizSet.description,
-                    shareNum = sharedQuizSet.quizSet.sharedQuizSets.count().toLong(),
-                    viewNum = sharedQuizSet.quizSet.views.count().toLong(),
+                    shareNum = sharedQuizSet.quizSet.sharedQuizSets.count(),
+                    viewNum = sharedQuizSet.quizSet.views.count(),
                     id = sharedQuizSet.quizSet.id!!
                 )
                 quizSetsFromMySharedQuizSets.add(mySharedQuizSet)
